+++ spek-ide-plugin/android-studio3.2/src/main/resources/META-INF/plugin.xml
@@ -24,6 +24,12 @@
         <jvmParameterPatcher implementation="org.spekframework.android.Spek2AndroidParameterPatcher"/>
     </extensions>
 
+    <project-components>
+        <component>
+            <implementation-class>org.spekframework.intellij.domain.ScopeDescriptorCache</implementation-class>
+        </component>
+    </project-components>
+
     <actions>
         <!-- Add your actions here -->
     </actions>
